Made default content loading configurable
The default content is now a launch parameter, and every single
asset is now loaded in using the virtual FS mechanism, bar from a few
implementation stuff you shouldn't mess with in the first place. 

This change allows finally support to bundle the base game content as a
.zip file, and opens way for the core content to be a seperate project
from the implementation altogether.