Added a task that generates a version.txt file
This file contains the verion specifed in Gradle, the commit id/hash and
the build date. It is of a different syntax than the usual version.txt,
and so the launcher will have to be updated to reflect this fancier
nature. This is being done in preperation for an automated release/staging
CI system.